requirements.txt
streamlit
langchain_openai
langchain_core
langchain_groq
gtts
SpeechRecognition
# speech_recognition
docx
python-dotenv
PyAudio